requirements.txt
numpy==1.23
pytest==7.1
python-dotenv==0.20
pytorch_lightning==1.7
tabulate==0.8.10
torch==1.12
transformers==4.10
pytest-cov==3.0